move ClassInfo to separate file
